% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaner.R
\name{clean_pdb}
\alias{clean_pdb}
\title{PDB Cleaning.}
\usage{
clean_pdb(pdb)
}
\arguments{
\item{pdb}{Input containing only the name of the 4-digit PDB file, the file will be obtained online. If there is an extension ".pdb" or full path, the file will be obtained locally.}
}
\description{
Cleaning PDB performs the reorganization of data from an input
PDB file. This function is responsible for removing lines and
information that are not necessary for calculating occluded
areas. In addition, the residues and their atoms are renumbered
to perform the area calculations.
}
\author{
Carlos Henrique da Silveira (carlos.silveira@unifei.edu.br)

Herson Hebert Mendes Soares (hersonhebert@hotmail.com)

João Paulo Roquim Romanelli (joaoromanelli@unifei.edu.br)

Patrick Fleming (Pat.Fleming@jhu.edu)
}
