% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders_manipulate.R
\name{create_folder}
\alias{create_folder}
\title{Files Manipulating.}
\usage{
create_folder(pdbname)
}
\description{
Function for creating folders and manipulating files in the FIBOS system.
}
\author{
Carlos Henrique da Silveira (carlos.silveira@unifei.edu.br)

Herson Hebert Mendes Soares (hersonhebert@hotmail.com)

João Paulo Roquim Romanelli (joaoromanelli@unifei.edu.br)

Patrick Fleming (Pat.Fleming@jhu.edu)
}
