% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_intermediary.R
\name{execute}
\alias{execute}
\alias{call_main}
\title{Surface Calc}
\usage{
call_main(iresf, iresl, maxres, maxat)
}
\arguments{
\item{iresf}{The number of the first element in the PDB}

\item{iresl}{The number of the first element in the PDB}

\item{maxres}{Maximum number of residues.}

\item{maxat}{Maximum number of atoms.}
}
\description{
The implemented function executes the implemented methods.
Using this function, it is possible to calculate occluded areas
through the traditional methodology, Occluded Surface, or by
applying the Fibonacci OS methodology. At the end of the method
execution, the "prot.srf" file is generated, and returned for
the function. The data in this file refers to all contacts
between atoms of molecules present in a protein's PDB.
}
\seealso{
\code{\link[=read_prot]{read_prot()}}

\code{\link[=read_osp]{read_osp()}}

\code{\link[=occluded_surface]{occluded_surface()}}
}
\author{
Carlos Henrique da Silveira (carlos.silveira@unifei.edu.br)

Herson Hebert Mendes Soares (hersonhebert@hotmail.com)

João Paulo Roquim Romanelli (joaoromanelli@unifei.edu.br)

Patrick Fleming (Pat.Fleming@jhu.edu)
}
