% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respak.R
\name{osp}
\alias{osp}
\title{Occluded Surface Packing (OSP)}
\usage{
osp(file)
}
\arguments{
\item{file}{a SRF File (.srf) generated by occluded_surface in fibos_files folder.}
}
\value{
A table containing:
\item{\code{Resnum}}{residue id.}
\item{\code{Resname}}{residue name.}
\item{\code{OS}}{the summed areas of dots in residue.}
\item{\code{`os*[1-raylen]`}}{OS areas weighted by (1-raylen). Raylen is the average lengths of normals normalized by 2.8 Å (water diameter). So, raylen is a value between 0 and 1. A raylen close to 1 indicates worse packaging, and the OS will be reduced.}
\item{\code{OSP}}{average occluded surface packing value (OSP) by residue.}
}
\description{
Implements the occluded surface packing density metric (OSP) averaged by residue, as described in (Fleming and Richards 2000).
}
\examples{
library(fibos)

# Calculate FIBOS per atom and create .srf files in fibos_files folder
pdb_fibos <- occluded_surface("1fib", method = "FIBOS")

# Calculate OSP metric per residue from .srf file in fibos_files folder
pdb_osp <- osp(fs::path("fibos_files","prot_1fib.srf"))

}
\references{
Fleming PJ, Richards FM. Protein packing: Dependence on protein size, secondary structure and amino acid composition. J Mol Biol 2000;299:487–98.(\url{https://doi.org/10.1006/jmbi.2000.3750})

Pattabiraman N, Ward KB, Fleming PJ. Occluded molecular surface: Analysis of protein packing. J Mol Recognit 1995;8:334–44. (\url{https://doi.org/10.1002/jmr.300080603})

Herson H. M. Soares, João P. R. Romanelli, Patrick J. Fleming, Carlos H. da Silveira. bioRxiv, 2024.11.01.621530. (\url{https://doi.org/10.1101/2024.11.01.621530})
}
\seealso{
\link{occluded_surface}
}
\author{
Herson Soares, João Romanelli, Patrick Fleming, Carlos Silveira.
}
