% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot_read.R
\name{read_prot}
\alias{read_prot}
\title{Prot Reading}
\usage{
read_prot(file)
}
\arguments{
\item{file}{Path or name of prot file.}
}
\description{
It is crucial to read and manipulate the data of the occluded
surfaces that were calculated. Through the results, it is
possible to draw important conclusions and perform studies on
specific proteins and their behavior at the atomic level.
}
\details{
The function allows reading a prot file, which contains data
calculated using the OS or FIBOS methodology. The input parameter
is the name of the file "prot.srf" or the directory path where it is
located."
}
\seealso{
\code{\link[=occluded_surface]{occluded_surface()}}

\code{\link[=osp]{osp()}}

\code{\link[=read_osp]{read_osp()}}
}
\author{
Carlos Henrique da Silveira (carlos.silveira@unifei.edu.br)

Herson Hebert Mendes Soares (hersonhebert@hotmail.com)

João Paulo Roquim Romanelli (joaoromanelli@unifei.edu.br)

Patrick Fleming (Pat.Fleming@jhu.edu)
}
