% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_os.R
\name{occluded_surface}
\alias{occluded_surface}
\title{Calcule the Surface}
\usage{
occluded_surface(pdb, method = "FIBOS", verbose = FALSE)
}
\arguments{
\item{pdb}{Input containing only the name of the 4-digit PDB file, the file will be obtained online. If there is an extension ".pdb" or full path, the file will be obtained locally.}

\item{method}{Method to be used: OS or FIBOS}
}
\description{
The calculation of occluded surface areas is essential for
understanding the possibility of an enzyme passing between atoms
of a protein. To perform the calculation, it is considered that
a surface is occluded based on tests with a probe, which is
typically the water molecule.
}
\seealso{
\code{\link[=read_prot]{read_prot()}}

\code{\link[=read_osp]{read_osp()}}

\code{\link[=osp]{osp()}}
}
\author{
Carlos Henrique da Silveira (carlos.silveira@unifei.edu.br)

Herson Hebert Mendes Soares (hersonhebert@hotmail.com)

João Paulo Roquim Romanelli (joaoromanelli@unifei.edu.br)

Patrick Fleming (Pat.Fleming@jhu.edu)
}
