% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respak.R
\name{osp}
\alias{osp}
\alias{read_osp}
\title{Respak Calcule}
\usage{
osp(file)

read_osp(prot_file)
}
\arguments{
\item{file}{Prot File (.srf).}

\item{prot_file}{OSP File (.pak).}
}
\description{
The OSP value is important for verifying the quality of the values
calculated by the developed package for calculating the contact
areas between the molecules of the analyzed protein.

The OSP value is important for verifying the quality of the values
calculated by the developed package for calculating the contact
areas between the molecules of the analyzed protein.
}
\seealso{
\code{\link[=read_prot]{read_prot()}}

\code{\link[=occluded_surface]{occluded_surface()}}

\code{\link[=read_osp]{read_osp()}}
}
\author{
Carlos Henrique da Silveira (carlos.silveira@unifei.edu.br)

Herson Hebert Mendes Soares (hersonhebert@hotmail.com)

João Paulo Roquim Romanelli (joaoromanelli@unifei.edu.br)

Patrick Fleming (Pat.Fleming@jhu.edu)
}
